requirements.txt
Cerberus==1.3.4
certifi==2022.6.15
charset-normalizer==2.1.0
docker==5.0.3
idna==3.3
podman==4.0.0
pyxdg==0.28
PyYAML==6.0
requests==2.28.1
toml==0.10.2
urllib3==1.26.11
websocket-client==1.3.3
